requirements.txt
mcp>=1.0.0
fastmcp>=0.2.0
httpx>=0.25.0
aiohttp>=3.8.0
python-dotenv>=1.0.0 